# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Nina Ofertas API",
  "version": "0.5.0"
 },
 "paths": {
...
  "/api/expenses": {
   "get": {
    "tags": [
     "expenses"
    ],
    "summary": "List Expenses",
    "operationId": "list_expenses_api_expenses_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "from",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "From"
      }
     },
     {
      "name": "to",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "To"
      }
     },
     {
      "name": "category_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Category Id"
      }
     },
     {
      "name": "campaign_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Campaign Id"
      }
     },
     {
      "name": "bot_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Bot Id"
      }
     },
     {
      "name": "platform_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Platform Id"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "-created_at",
       "title": "Sort"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PaginatedResponse_ExpenseResponse_"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "expenses"
    ],
    "summary": "Create Expense",
    "operationId": "create_expense_api_expenses_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ExpenseCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ExpenseResponse"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   }
  },
  "/api/expenses/{expense_id}": {
   "patch": {
    "tags": [
     "expenses"
    ],
    "summary": "Update Expense",
    "operationId": "update_expense_api_expenses__expense_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "expense_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Expense Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ExpenseUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ExpenseResponse"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "expenses"
    ],
    "summary": "Delete Expense",
    "operationId": "delete_expense_api_expenses__expense_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "expense_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Expense Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   }
  },
  "/api/expense-categories": {
   "get": {
    "tags": [
     "expenses"
    ],
    "summary": "List Categories",
    "operationId": "list_categories_api_expense_categories_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ExpenseCategoryResponse"
         },
         "type": "array",
         "title": "Response List Categories Api Expense Categories Get"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "expenses"
    ],
    "summary": "Create Category",
    "operationId": "create_category_api_expense_categories_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ExpenseCategoryCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ExpenseCategoryResponse"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/expenses/import": {
   "post": {
    "tags": [
     "expenses"
    ],
    "summary": "Import Expenses",
    "operationId": "import_expenses_api_expenses_import_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_import_expenses_api_expenses_import_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ImportResult"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_import_expenses_api_expenses_import_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_import_expenses_api_expenses_import_post"
   },
...
   "ErrorEnvelope": {
    "properties": {
     "error": {
      "$ref": "#/components/schemas/ErrorDetail"
     }
    },
    "type": "object",
    "required": [
     "error"
    ],
    "title": "ErrorEnvelope"
   },
...
   "ExpenseCategoryCreate": {
    "properties": {
     "slug": {
      "type": "string",
      "maxLength": 80,
      "minLength": 1,
      "pattern": "^[a-z0-9][a-z0-9_-]*$",
      "title": "Slug"
     },
     "name": {
      "type": "string",
      "maxLength": 120,
      "minLength": 1,
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "slug",
     "name"
    ],
    "title": "ExpenseCategoryCreate"
   },
   "ExpenseCategoryResponse": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "name": {
      "type": "string",
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "id",
     "slug",
     "name"
    ],
    "title": "ExpenseCategoryResponse"
   },
   "ExpenseCreate": {
    "properties": {
     "description": {
      "type": "string",
      "maxLength": 500,
      "minLength": 1,
      "title": "Description"
     },
     "amount": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       }
      ],
      "title": "Amount",
      "decimal_places": 2,
      "ge": "0",
      "max_digits": 14
     },
     "incurred_on": {
      "type": "string",
      "format": "date",
      "title": "Incurred On"
     },
     "currency": {
      "type": "string",
      "maxLength": 3,
      "minLength": 3,
      "title": "Currency",
      "default": "BRL"
     },
     "category_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category Id"
     },
     "platform_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Platform Id"
     },
     "campaign_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Campaign Id"
     },
     "bot_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bot Id"
     },
     "niche_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Niche Id"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "description",
     "amount",
     "incurred_on"
    ],
    "title": "ExpenseCreate"
   },
   "ExpenseResponse": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "description": {
      "type": "string",
      "title": "Description"
     },
     "amount": {
      "type": "string",
      "title": "Amount"
     },
     "currency": {
      "type": "string",
      "title": "Currency"
     },
     "incurred_on": {
      "type": "string",
      "format": "date",
      "title": "Incurred On"
     },
     "category_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category Id"
     },
     "platform_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Platform Id"
     },
     "campaign_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Campaign Id"
     },
     "bot_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bot Id"
     },
     "niche_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Niche Id"
     },
     "source": {
      "type": "string",
      "title": "Source"
     },
     "external_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "External Id"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "description",
     "amount",
     "currency",
     "incurred_on",
     "category_id",
     "platform_id",
     "campaign_id",
     "bot_id",
     "niche_id",
     "source",
     "external_id",
     "notes",
     "created_at"
    ],
    "title": "ExpenseResponse"
   },
...
   "ImportErrorItem": {
    "properties": {
     "line": {
      "type": "integer",
      "title": "Line"
     },
     "message": {
      "type": "string",
      "title": "Message"
     }
    },
    "type": "object",
    "required": [
     "line",
     "message"
    ],
    "title": "ImportErrorItem"
   },
   "ImportResult": {
    "properties": {
     "imported": {
      "type": "integer",
      "title": "Imported"
     },
     "skipped": {
      "type": "integer",
      "title": "Skipped"
     },
     "errors": {
      "items": {
       "$ref": "#/components/schemas/ImportErrorItem"
      },
      "type": "array",
      "title": "Errors"
     }
    },
    "type": "object",
    "required": [
     "imported",
     "skipped",
     "errors"
    ],
    "title": "ImportResult"
   },
...
   "PaginatedResponse_ExpenseResponse_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ExpenseResponse"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "PaginatedResponse[ExpenseResponse]"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}